+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## UNRELEASED
 
+* Expose unsafe constructors `UnsafeMkId` and `UnsafeMkArg` for `Id` and `Arg`
+  respectively from `Covenant.Test`
+
 ## 1.3.0 -- 2025-10-07
 
 * Zipper for the ASG in `Covenant.Zipper`